Add files to make project work with Google Cloud Platform.
- app.yaml: Project metadata that GCP needs to be aware of.
- appengine_config.py: A python file which specifies a directory
(possibly more than one) where other third-party python libraries are
located (the libraries you'd install from pip, or something like that).
- requirements.txt: A list of 3rd party packages being used. To create
this, run the command `pip freeze` and place the output from that
command in this file. This isn't strictly necessary, but it makes
keeping which packages are installed easy to keep track of.

Modified the .gitignore file to ignore the directory `lib`, which is
necessary for GCP's third party libraries, but not necessary anywhere
else.

Changes to be committed:
      new file:   app.yaml
      new file:   appengine_config.py
      new file:   requirements.txt
      modified:   .gitignore