Change function definitions to be MariaDB/binary logging compatible.
The Google Cloud Platform's database is MariaDB, so the definitions of
the functions and procedures must be made compatible with it.

It seems that MariaDB has binary logging on by default, which
requires these little hints about what functions do, so that MariaDB can
figure out some extra things.

Changes to be committed:
      modified:   calc_base_fare.sql
      modified:   calc_full_fare.sql
      modified:   create_trip_stations_v2.sql
      modified:   drop-tables.sql
      modified:   free_seat_check.sql
      modified:   free_seat_decrement.sql
      modified:   free_seat_increment.sql
      modified:   generate-seats.sql
      modified:   show_trains_v3.sql
      modified:   trip_res_cancel_v2.sql