Make our project ready for deployment on GCP.
The changes to `app.py`, `app.yaml`, and `config.py` are about providing our
project with the appropriate metadata for connecting to the GCP SQL
instance from wherever it is deployed.

The python package `gunicorn` was added to our project because GCP uses
that to get web pages from our application. `requirements.txt` was
changed to reflect this.

The templates were changed so that all links to external stylesheets
used `https` instead of `http`. When the browser tries to load these
from an actual website on the internet, it may refuse to do so, seeing
the links as unsafe.

Changes to be committed:
      modified:   app.py
      modified:   app.yaml
      new file:   config.py
      modified:   requirements.txt
      modified:   templates/index.html
      modified:   templates/login.html
      modified:   templates/register.html
      modified:   templates/results.html