Change app.py to be able to connect to linux lab database on host.
Now, app.py first attempts to connect to the linux lab database before
trying to connect to a local database. It also notifies the user if it
can't, and shows why it couldn't connect to either database.

Creating a connection is now handled by the create_db function.

Furthermore, the database name has been changed to S17336team1, to match
the name of the database in the linux lab. If anyone has a local
database named `s17336team1` and wants to run the web server with a
local database, they'll have to change the database name to match.

Changes to be committed:
       modified:   .gitignore
       modified:   app.py
       new file:   requirements.txt