Fix locking mechanism in front-end.
Turns out that when a session locks tables, the session can only access
the tables which it locked. So I had to change the locking statement to
lock every table touched by the create_trips function.

The changes have been tested locally on my computer, though the
environment is very similar to the actual server.

Changes to be committed:
      modified:   app.py